convert the bugsnag-cli-install.js to ts
